backend/requirements.txt
fastapi
uvicorn
bertopic
arxiv
pymed
scikit-learn
pandas
numpy
pydantic
python-multipart
sqlalchemy
apscheduler
httpx
python-dateutil
sentence-transformers
umap-learn
hdbscan
openai
ruptures
statsmodels
networkx
python-louvain
